#19 Order and payment flow, fixing hmmm

+++ wallet/src/main/java/com/chuadatten/wallet/entity/WalletReservation.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -30,7 +31,7 @@ public class WalletReservation {
     private UUID orderId;
 
     @Column(name = "amount", nullable = false)
-    private Long amount;
+    private BigInteger amount;
 
     @Column(name = "currency", nullable = false, columnDefinition = "CHAR(3)")
     private String currency;
